rename state and modules files to uyuni_config.py
