+++ susemanager-utils/susemanager-sls/src/modules/uyuni_config.py
