+++ susemanager-utils/susemanager-sls/src/states/uyuni_config.py
